flatten module structure
