+++ dev/v1.ex
@@ -2,9 +2,9 @@ defmodule Inngest.Dev.EventFn2 do
   @moduledoc false
 
   use Inngest.Function
-  alias Inngest.Function.{Opts, Trigger}
+  alias Inngest.{FnOpts, Trigger}
 
-  @func %Opts{id: "test-func-v2", name: "test func v2"}
+  @func %FnOpts{id: "test-func-v2", name: "test func v2"}
   @trigger %Trigger{event: "test/hello"}
 
   @impl true
