--- lib/inngest/function/opts.ex
@@ -1,15 +1,0 @@
-defmodule Inngest.Function.Opts do
-  @moduledoc false
-
-  defstruct [
-    :id,
-    :name,
-    :retries
-  ]
-
-  @type t() :: %__MODULE__{
-          id: binary(),
-          name: binary(),
-          retries: number()
-        }
-end
